+++ app/main.py
@@ -5,9 +5,9 @@ from fastapi import FastAPI
 from tortoise import Tortoise
 from tortoise.exceptions import DBConnectionError
 
-from app.ai.router import router as ai_router
-from app.tag.router import router as tag_router
-from app.user.routes import router as user_router  # 유저 라우터 추가
+from app.ai.api import router as ai_router
+from app.tag.api import router as tag_router
+from app.user.api import router as user_router  # 유저 라우터 추가
 from core.config import TORTOISE_ORM
 
 DATABASE_URL = "postgresql+asyncpg://diaryapi:diaryapi@localhost:5432/diaryapi"
@@ -60,7 +60,7 @@ app = FastAPI(
 # AI 라우터 등록
 app.include_router(ai_router)
 app.include_router(tag_router)
-app.include_router(user_router, prefix="/users", tags=["Users"])  # 유저 라우터 추가
+app.include_router(user_router)  # 유저 라우터 추가
 
 
 @app.get("/")
