+++ app/notification/schema.py
@@ -7,16 +7,13 @@ from app.notification.model import NotificationType
 
 class NotificationCreateRequest(BaseModel):
     content: str = Field(..., max_length=255, description="알림 내용")
-    notification_type: NotificationType = Field(..., description="알림 타입")
-    # TODO : user schema에 receive_notifications 필드 추가?
-    # TODO : user schema에 알림 수신 여부 설정 변경 스키마 추가
-    # TODO : user api에 알림 수신 여부 설정 변경 엔드포인트 추가
+    type: NotificationType = Field(..., description="알림 타입")
 
 
 class NotificationResponse(BaseModel):
-    notification_id: int
+    id: int
     content: str
-    notification_type: NotificationType
+    type: NotificationType
     created_at: datetime
     updated_at: datetime
